+++ apps/workspace-host/src/lib/aws.ts
@@ -1,27 +1,30 @@
-import { type S3ClientConfig } from '@aws-sdk/client-s3';
+import { fromNodeProviderChain } from '@aws-sdk/credential-providers';
+import { makeAwsConfigProvider } from '@prairielearn/aws';
+
 import { config } from './config';
 
-export function makeS3ClientConfig(extraConfig: S3ClientConfig = {}): S3ClientConfig {
-  const newConfig = makeAwsClientConfig(extraConfig);
-
-  if (!config.runningInEc2) {
-    // If we're not running in EC2, assume we're running with a local s3rver instance.
-    // See https://github.com/jamhall/s3rver for more details.
-    newConfig.forcePathStyle = true;
-    newConfig.credentials = {
-      accessKeyId: 'S3RVER',
-      secretAccessKey: 'S3RVER',
-    };
-    newConfig.endpoint = 'http://localhost:5000';
-  }
-
-  return newConfig;
-}
-
-export function makeAwsClientConfig<T extends Record<string, any> = object>(extraConfig = {} as T) {
-  return {
+const awsConfigProvider = makeAwsConfigProvider({
+  credentials: fromNodeProviderChain(),
+  getClientConfig: () => ({
     region: config.awsRegion,
-    endpoint: process.env.AWS_ENDPOINT,
-    ...extraConfig,
-  };
-}
+  }),
+  getS3ClientConfig: () => {
+    if (!config.runningInEc2) {
+      // If we're not running in EC2, assume we're running with a local s3rver instance.
+      // See https://github.com/jamhall/s3rver for more details.
+      return {
+        forcePathStyle: true,
+        credentials: {
+          accessKeyId: 'S3RVER',
+          secretAccessKey: 'S3RVER',
+        },
+        endpoint: 'http://localhost:5000',
+      };
+    }
+
+    return {};
+  },
+});
+
+export const makeAwsClientConfig = awsConfigProvider.makeAwsClientConfig;
+export const makeS3ClientConfig = awsConfigProvider.makeS3ClientConfig;