+++ packages/postgres/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @prairielearn/postgres
 
+## 1.7.7
+
+### Patch Changes
+
+- bd0053577: Upgrade all dependencies
+
+## 1.7.6
+
+### Patch Changes
+
+- f03853d90: Upgrade all dependencies
+
+## 1.7.5
+
+### Patch Changes
+
+- 6cad75197: Upgrade all dependencies
+
 ## 1.7.4
 
 ### Patch Changes