+++ packages/sanitize/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @prairielearn/sanitize
 
+## 1.0.7
+
+### Patch Changes
+
+- bd0053577: Upgrade all dependencies
+
+## 1.0.6
+
+### Patch Changes
+
+- f03853d90: Upgrade all dependencies
+
+## 1.0.5
+
+### Patch Changes
+
+- 6cad75197: Upgrade all dependencies
+
 ## 1.0.4
 
 ### Patch Changes